Fix the remaining parser crashes, and prove the arithmetic under them
Completes the eight fuzzer findings from the previous commit and adds a third
verification layer over the same code.

Parser fixes (wasm, dex, apk, jvm, jar, pdb; pyc and lua landed earlier). All
nine fuzz targets now replay clean, and 42 standalone suites pass. Every fix
carries a regression test that was confirmed to fail against the unfixed code.
The PDB work also contributes a hand-built corpus of malformed inputs -- a
stream count that overflows, a root page outside the file, an unterminated
module-name table -- so those paths stay covered by name rather than by
whatever the mutator happens to reach.

The proofs. Fuzzing shows a bug exists; it cannot show that one does not. Every
crash found in this tree was in the same few lines of arithmetic: a count,
length or offset read out of a file and used before anything checked the input
could supply it. That arithmetic now lives in two header-only, dependency-free
files -- retdec/utils/bounds.h and retdec/utils/leb128.h -- and
scripts/verify_esbmc.sh proves them with ESBMC and an SMT solver. 39 properties,
over the whole 64-bit range, not for sampled values. The harnesses have no
unbounded loops, so these are proofs rather than bounded searches.

Callers use the headers instead of re-deriving the logic, which is what makes
the proofs mean anything: mini_emu for page and offset arithmetic,
eh_reconstruct for LEB128 accumulation and sign extension.

That migration fixed a live bug. IBinaryView::readSLEB128 accumulated into an
int64_t with no bound on the shift. Both halves are undefined behaviour on
attacker-controlled DWARF -- (int64_t)0x7F << 57 overflows the signed range, and
a run of continuation bytes drives the shift past 64 -- and the loop had no byte
limit, relying on readU8 returning 0 past the end of the view to stop it.

Writing the proofs refuted four claims:

* countFits accepted a zero count from a position past the end of the buffer,
  because remaining() saturates. That broke composition with rangeFits.
  Counterexample: pos = SIZE_MAX - 34, size = 0, count = 0.
* pageCount(...) * pageSize was asserted representable. It is not: bytes =
  SIZE_MAX with pageSize = 2 gives a count of 2^63 whose product wraps. The
  header documents the contract; the proof states coverage by division.
* The LEB128 accumulator shifted a byte's payload past the top of the word.
  Defined for an unsigned type, but dropping those bits silently is how an
  over-long encoding decodes as a plausible value. The payload is now masked to
  the bits that fit, so the truncation is explicit.
* Two harnesses committed the fault they were proving absent -- asserting
  `a + b < a` to show a sum wraps, and building a symbolic value with
  `v = (v << 8) | byte`. A proof that starts by discarding information is not a
  proof.

Limits are stated in docs/VERIFICATION.md rather than glossed. The parsers
themselves are not verified: ESBMC's models of the C++ standard library do not
stretch to this codebase (std::variant caps at four alternatives, std::min has
no initializer_list overload). And a query holding both a symbolic 64-bit
multiply and a symbolic 64-bit divide is nonlinear bitvector arithmetic that no
solver here discharges, so those properties are proved once per element width
that actually occurs -- position, size and count stay fully symbolic throughout.

tests/bounds/ adds runtime tests for both headers, carrying the DWARF standard's
LEB128 vectors. Proofs cover safety for all inputs but say nothing about the
values a decoder produces; the two answer different questions.

+++ src/pyc_parser/pyc_reader.cpp
@@ -569,11 +569,11 @@ std::vector<uint32_t> PycReader::findLeaders(const PyCodeObject& code) const {
     // prefixes it chains -- a compiler emits at most three, a hostile file as
     // many as it likes -- so the accumulator has to wrap at that width instead
     // of shifting a signed value out of range.
-    int32_t extArg = 0;
+    uint32_t extArg = 0;
 
     while (pos < bytecode.size()) {
         uint8_t op = bytecode[pos];
-        int32_t  arg = 0;
+        uint32_t arg = 0;
         size_t   instrSize = 2; // always 2 bytes in wordcode
 
         if (pos + 1 < bytecode.size())
@@ -662,12 +662,12 @@ void PycReader::buildCFG(const PyCodeObject& code, BcMethod& method) const {
     const bool is311 = code.version.atLeast(3, 11);
     const uint8_t haveArg = haveArgument(code.version);
     size_t pos   = 0;
-    int32_t extArg = 0;
+    uint32_t extArg = 0;  // unsigned for the reason given in findLeaders
     size_t   curBlockIdx = 0;
 
     while (pos < bytecode.size()) {
         uint8_t op = bytecode[pos];
-        int32_t  arg = (pos + 1 < bytecode.size()) ? bytecode[pos+1] : 0;
+        uint32_t arg = (pos + 1 < bytecode.size()) ? bytecode[pos+1] : 0;
 
         // Check if we've entered a new block
         {
@@ -685,7 +685,7 @@ void PycReader::buildCFG(const PyCodeObject& code, BcMethod& method) const {
             continue;
         }
 
-        int32_t fullArg = (extArg | arg);
+        int32_t fullArg = static_cast<int32_t>(extArg | arg);
         extArg = 0;
 
         BcInstruction insn;
